// src/layout/pages/switches/Switches.interface.ts
export interface ISwitch {
  category: string;
  createdAt: string;
  switchDesc: string;
  switchDiscountRate: string;
  switchFeatures: {
    color: string;
  };
  switchId: string;
  switchImagePath: string;
  switchImageUrl: string;
  switchName: string;
  switchPrice: string;
  switchStock: string;
  updatedAt: string;
}